+++ niftynet/engine/handler_checkpoint.py
