+++ niftynet/engine/handler_console.py
