+++ niftynet/engine/handler_gradient.py
